// references/math-raiders/client/src/spacetime/operation_type.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
import {
  TypeBuilder as __TypeBuilder,
  t as __t,
  type AlgebraicTypeType as __AlgebraicTypeType,
  type Infer as __Infer,
} from "spacetimedb";

// The tagged union or sum type for the algebraic type `Operation`.
const Operation = __t.enum("Operation", {Add: __t.unit(),
  Subtract: __t.unit(),
  Multiply: __t.unit(),
  Divide: __t.unit(),
});

export default Operation;